Updates instructions for setup
As a Python newbie, I changed the order of operations for project  setup that made more sense to me, namely activating the virtual environment before proceeding with `pip install` calls. I'm guessing that installs the dependencies in my local virtual environment rather than my computer's global install site, which is what I want.  Also, added the `.txt` to the install command, as it seemed to be required on my computer.