Adds very basic functionality to diff step output
Adds a switch (`--diff`) to use in running a pipeline through the
command line app. When in use, the changes that are made to each row are
output to stdout after each step.

This does not capture rows that are dropped or added and is only
implemented for `@row_step`s.