Properly preserves row number during batch steps
This is not fool-proof, as one of the test cases demonstrates, but if
the user mutates rows in place during a batch step, rather than creating
them from scratch, then we can easily preserve row numbers.

If the user creates new dictionaries to return, then it is going to be
a complex process for us to try to figure out which row the data
originated from. We could have identifying keys, or we could try to find
the most closely matching row and use a heuristic. Or we could make the
row_num data in the PhaseRecord a known part of the API and rely on the
user to code up batch steps that preserve it explicitly.