Supports defining extra_sources as IOObjects
Allows for setting extra sources as IOObjects or as strings. Strings are
still valuable for when the source will be piped in from another phase.
(Perhaps that should be specified as an explicit type of IOObject to be
really clear about it)

Removes the ReadWriteObject in preference for IOObject for managing
extra sources and outputs

ExtraMapping and ExtraRecords no longer subclass UserDict and UserList,
and the data from them is explicitly passed through to the steps. This
enables the use of `defaultdict` and the like.